Modify write_page, loop_pages methods
The write_page method tests that a line from the story file meets a certain criteria (is or is not a Page or Options header) and passes line into loop_pages method. The loop_pages method adds the line to the Text array if the line is not an Options header, a Page header, or an option. If it is an Options header, it prints an alert to the console instead of inserting the line into the Text array. If the line is an option, it writes the option text and the page number into a tuple in the Options array.